nightly test for inverse analysis is activated again
changed input parameter INV_ANA_MAX_RUN from double
to integer



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@10527 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_inv_analysis/inv_analysis.cpp
@@ -134,7 +134,7 @@ void STR::InvAnalysis::Integrate()
   const Teuchos::ParameterList& iap = DRT::Problem::Instance()->InverseAnalysisParams();
   double alpha  = iap.get<double>("INV_ALPHA");
   double beta   = iap.get<double>("INV_BETA");
-  int max_itter = iap.get<double>("INV_ANA_MAX_RUN");
+  int max_itter = iap.get<int>("INV_ANA_MAX_RUN");
   output_->NewResultFile((numb_run_-1));
   // fitting loop
   do
